mrc: refactor interactive mode into its own module
